pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Sleptsov Denis Sergeevich <hurrdurrrderp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.2.1"
fastapi = {extras = ["standard-no-fastapi-cloud-cli"], version = "^0.116.0"}
sqlmodel = "^0.0.24"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"
bandit = "^1.7.8"
pytest = "^8.1.1"
pre-commit = "^3.7.0"
flake8-pyproject = "^1.2.3"
types-pyyaml = "^6.0.12.20240311"

[tool.bandit]
exclude_dirs = [
    "__pycache__",
    ".mypy_cache",
    ".venv"
]

[tool.black]
line-length = 100
skip-string-normalization = true

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".mypy_cache",
    ".venv"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
